Day 9: We have a level! And music! SO close now
need to soften the edges of all the inner levels, give edge material to the maze, make the pool balls, and spin to the turret floor.
ah, and finish the music, make the logo, record the intro, make the intro slide, put texts of the names of the areas, widen the camera (specially up and down), reset the camera to the front of the model, enable the grapple hook, make menus.